rename Util class to Args
Made a copy of the class to be backward compatible.
