+++ vars/executable.groovy
@@ -1,6 +1,6 @@
 import com.puzzleitc.jenkins.command.ExecutableCommand
 import com.puzzleitc.jenkins.command.context.JenkinsPipelineContext
-import static com.puzzleitc.jenkins.Util.parseArgs
+import static com.puzzleitc.jenkins.util.Args.parseArgs
 
 def call(Map namedArgs = [:], Object... positionalArgs) {
     def args = parseArgs(namedArgs, positionalArgs, ["name"], ["toolName": null])
