+++ vars/kustomize.groovy
@@ -1,7 +1,7 @@
 import com.puzzleitc.jenkins.command.KustomizeCommand
 import com.puzzleitc.jenkins.command.context.JenkinsPipelineContext
 
-import static com.puzzleitc.jenkins.Util.parseArgs
+import static com.puzzleitc.jenkins.util.Args.parseArgs
 
 def call(Map namedArgs = [:], Object... positionalArgs) {
     def args = parseArgs(namedArgs, positionalArgs, ['path'])
