+++ vars/withGitCrypt.groovy
@@ -1,6 +1,6 @@
 import com.puzzleitc.jenkins.command.GitCryptCommand
 import com.puzzleitc.jenkins.command.context.JenkinsPipelineContext
-import static com.puzzleitc.jenkins.Util.parseArgs
+import static com.puzzleitc.jenkins.util.Args.parseArgs
 
 def call(Map namedArgs = [:], Object... positionalArgs) {
     def args = parseArgs(namedArgs, positionalArgs, ['credentialsId', 'body'])
